pyproject.toml
[tool.poetry]
name = "mnist_recognition"
version = "0.1.0"
description = "Framework for object_detection"
authors = ["MadDrive <letisaudep@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
numpy = "^1.21.2"
pandas = "^1.3.3"
opencv-python = "^4.5.5"
scikit-image = "^0.19.1"
ipywidgets = "^7.6.5"
imutils = "^0.5.4"
pyzmq = "~=19.0.2"
pytest = "^7.0.1"
torch = "^1.11.0"
torchvision = "^0.12.0"
tensorboard = "^2.8.0"
scikit-learn = "^1.0.2"
scipy = "1.6.1"
albumentations = "^1.1.0"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.4.1"
black = "^21.9b0"
pre-commit = "^2.15.0"
autoflake = "^1.4"
pylint = "^2.11.1"
tqdm = "^4.62.3"
matplotlib = "^3.4.3"
seaborn = "^0.11.2"
nbdime = "^3.1.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://github.com/psf/black
target-version = ["py38"]
line-length = 80
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''
